% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropicTest.R
\name{entropicTest}
\alias{entropicTest}
\title{Function that applies of the hypothesis test for two time series,
a given embedding dimension, and
a given entropy ent = S (Shannon), T (Tsallis, beta is required),
R (Rényi, beta is required), or F (Fisher information measure)}
\usage{
entropicTest(TSx, TSy, emb, ent, beta, pvalue = 0.05)
}
\arguments{
\item{TSx}{a time series}

\item{TSy}{another time series}

\item{emb}{the embedding dimension}

\item{ent}{the type of entropy: S (Shannon), T (Tsallis, \eqn{\beta} is required), R (Rényi, \eqn{\beta} is required), or F (Fisher)}

\item{beta}{the beta parameter, only required for Tsallis and Rényi entropies}

\item{pvalue}{(optional, default 0.05) the \eqn{p}-value}
}
\description{
Function that applies of the hypothesis test for two time series,
a given embedding dimension, and
a given entropy ent = S (Shannon), T (Tsallis, beta is required),
R (Rényi, beta is required), or F (Fisher information measure)
}
