% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HFisher.R
\name{HFisher}
\alias{HFisher}
\title{Function that computes the normalized Fisher information measure of a probability function}
\usage{
HFisher(p)
}
\arguments{
\item{p}{a probability function}
}
\value{
a value in \eqn{[0,1]}
}
\description{
Given a probability function \eqn{\bm p=(p_1,p_2,\dots,p_k)},
its Fisher information measure is
\deqn{H^{(F)}[\bm p] = 4  \sum_{i=1}^{k-1} (\sqrt{p_{i+1}} - \sqrt{p_i})^2.}
}
