% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRenyi.R
\name{HRenyi}
\alias{HRenyi}
\title{Function that computes normalized Rényi entropy of order beta}
\usage{
HRenyi(p, beta)
}
\arguments{
\item{p}{a probability function}

\item{beta}{the parameter beta for the Rényi entropy; default 1.5}
}
\value{
a value in \eqn{[0,1]}
}
\description{
Given a probability function \eqn{\bm p=(p_1,p_2,\dots,p_k)}, its Rényi entropy of order \eqn{\beta \in \mathbb{R}^+\setminus\{1\}} is
\deqn{S^{(R_{\beta})}[\bm p] = \frac{1}{1-\beta} \log \sum_{i=1}^k p_i^\beta.}
We adopt the convention \eqn{0\ln 0 = 0}.
}
