% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HShannon.R
\name{HShannon}
\alias{HShannon}
\title{Function that computes the normalized Shannon entropy (using natural logarithms) of a probability function}
\usage{
HShannon(p)
}
\arguments{
\item{p}{a probability function}
}
\value{
A value in \eqn{[0,1]}
}
\description{
Given a probability function \eqn{\bm p=(p_1,p_2,\dots,p_k)},
its normalized Shannon entropy is
\deqn{S^{(S)}[\bm p] = -\frac{1}{\ln k}\sum_{i=1}^k p_i \ln p_i.}
We adopt the convention \eqn{0\ln 0 = 0}.
}
