% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPprob.R
\name{OPprob}
\alias{OPprob}
\title{Function that finds the Ordinal Patterns probabilities in a time series for a given embedding dimension}
\usage{
OPprob(TS, emb)
}
\arguments{
\item{TS}{time series of length \eqn{n-m+1}}

\item{emb}{embedding dimension \eqn{m}}
}
\value{
a sequence of \eqn{n} patterns
}
\description{
Function that finds the Ordinal Patterns probabilities in a time series for a given embedding dimension
}
