% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPseq.R, R/OPseqM2.R
\name{OPseq}
\alias{OPseq}
\title{Function that finds the OP sequence of a time series, given the
embedding dimension and lag}
\usage{
OPseq(TS, emb, lag = 1)

OPseq(TS, emb, lag = 1)
}
\arguments{
\item{TS}{time series}

\item{emb}{embedding dimension}

\item{lag}{time lag (default value: 1)}
}
\description{
Function that finds the OP sequence of a time series, given the
embedding dimension and lag

Function that finds the OP sequence of a time series, given the
embedding dimension and lag
}
