% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalM.R
\name{pvalM}
\alias{pvalM}
\title{Function that computes the p-value of the hypothesis test for two time series}
\usage{
pvalM(TSx, TSy, emb, ent, beta)
}
\arguments{
\item{TSx}{a time series}

\item{TSy}{another time series}

\item{emb}{embedding dimension}

\item{ent}{type of entropy: S (Shannon), T (Tsallis, beta is required), R (Rényi, beta is required), or F (Fisher)}

\item{beta}{parameter for the Tsallis and Rényi entropy}
}
\description{
Function that computes the p-value of the hypothesis test for two time series
}
