% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma2qM2.R
\name{sigma2qM2}
\alias{sigma2qM2}
\title{Function that computes the asymptotic variance of ordinal patterns' entropies}
\usage{
sigma2qM2(TS, emb, ent, beta)
}
\arguments{
\item{TS}{time series}

\item{emb}{embedding dimension}

\item{ent}{type of entropy: S (Shannon), T (Tsallis, \eqn{beta} is required), R (Rényi, \eqn{beta} is required), or F (Fisher)}

\item{beta}{the parameter for the Tsallis and Rényi entropies}
}
\value{
The asymptotic variance
}
\description{
Function that computes the asymptotic variance of ordinal patterns' entropies
}
