% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigmaM.R
\name{SigmaM}
\alias{SigmaM}
\title{Function that computes the covariance matrix of ordinal patterns
from a time series, given embedding dimension}
\usage{
SigmaM(TS, emb)
}
\arguments{
\item{TS}{time series}

\item{emb}{embedding dimension}
}
\value{
A covariance matrix
}
\description{
Function that computes the covariance matrix of ordinal patterns
from a time series, given embedding dimension
}
