% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigmaM2.R
\name{SigmaM2}
\alias{SigmaM2}
\title{Function that computes the covariance matrix of ordinal patterns
from a time series, given embedding dimension}
\usage{
SigmaM2(TS, emb)
}
\arguments{
\item{TS}{time series}

\item{emb}{embedding dimension}
}
\value{
A covariance matrix
}
\description{
Function that computes the covariance matrix of ordinal patterns
from a time series, given embedding dimension
}
