% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigmaqM2.R
\name{SigmaqM2}
\alias{SigmaqM2}
\title{Function that computes the covariance matrix for a time series}
\usage{
SigmaqM2(TS, emb, ent, beta)
}
\arguments{
\item{TS}{time series}

\item{emb}{embedding dimension}

\item{ent}{type of entropy: S (Shannon), T (Tsallis, \eqn{beta} is required), R (Rényi, \eqn{beta} is required), or F (Fisher)}

\item{beta}{parameter for the Tsallis and Rényi entropies}
}
\description{
Function that computes the covariance matrix for a time series
}
